+++ versioned/storage/store/src/test/java/org/projectnessie/versioned/storage/versionstore/TestVersionStoreReferenceCaching.java
@@ -53,6 +53,7 @@ public class TestVersionStoreReferenceCaching {
                 .clockNanos(clockNanos)
                 .referenceTtl(Duration.ofMinutes(1))
                 .referenceNegativeTtl(Duration.ofSeconds(1))
+                .cacheCapacityOvershoot(0.1d)
                 .build())
         .wrap(persist);
   }